refactor: move Table references to Backup out of base class
